requirements.txt
gitpython==3.1.27
bottle==0.12.23
python-magic==0.4.25